feat: ViewModel Factory 생성 로직을 각 ViewModel의 companion object로 분리하여 캡슐화
