+++ app/src/main/java/woowacourse/shopping/ui/shopping/ShoppingActivity.kt
@@ -9,8 +9,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.ui.Modifier
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.compose.viewModel
 import woowacourse.shopping.ShoppingApplication
 import woowacourse.shopping.data.remote.NetworkMonitor
@@ -33,17 +31,11 @@ class ShoppingActivity : ComponentActivity() {
                 Scaffold(modifier = Modifier.fillMaxSize()) { innerPadding ->
                     val viewModel: ShoppingViewModel =
                         viewModel(
-                            factory =
-                                object : ViewModelProvider.Factory {
-                                    override fun <T : ViewModel> create(modelClass: Class<T>): T =
-                                        ShoppingViewModel(
-                                            networkMonitor = NetworkMonitor(applicationContext),
-                                            productRepo = container.productRepository,
-                                            cartRepo = container.cartRepository,
-                                            recentProductRepo = container.recentProductRepository,
-                                            loadSize = loadSize,
-                                        ) as T
-                                },
+                            factory = ShoppingViewModel.provideFactory(
+                                container,
+                                NetworkMonitor(applicationContext),
+                                loadSize,
+                            ),
                         )
 
                     ShoppingScreen(
